[refactor] filterInput -> filter modal button으로 변경 - filterInput -> filter modal button으로 변경
