+++ client/src/components/molecules/FilterModal/FilterModalButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FilterModalButton';
